# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zeptly Social API",
    "version": "1.0.0",
    "description": "Provider-neutral social infrastructure gateway for Zeptly. All objects are canonical Zeptly Social objects; provider identifiers never appear. Authenticated with ZS1-HMAC-SHA256 service signatures (see docs/SECURITY.md)."
  },
  "components": {
    "securitySchemes": {
      "zeptlyServiceSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Zeptly-Signature",
        "description": "ZS1-HMAC-SHA256 signature. Also requires X-Zeptly-Caller, X-Zeptly-Timestamp and (workspace routes) X-Zeptly-Workspace-Id. See docs/SECURITY.md."
      }
    },
    "schemas": {
...
      "CreatePostRequestInput": {
        "type": "object",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/SocialContentInput"
          },
          "targets": {
            "minItems": 1,
            "maxItems": 25,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostTargetRequestInput"
            }
          },
          "externalRef": {
            "description": "Zeptly-side reference (e.g. content item id) for traceability",
            "type": "string",
            "maxLength": 256
          }
        },
        "required": [
          "content",
          "targets"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "AUTHENTICATION_FAILED",
                  "WORKSPACE_FORBIDDEN",
                  "NOT_FOUND",
                  "CONNECTION_NOT_FOUND",
                  "POST_NOT_FOUND",
                  "CONVERSATION_NOT_FOUND",
                  "MEDIA_NOT_FOUND",
                  "PROVISIONING_NOT_FOUND",
                  "PROVISIONING_EXPIRED",
                  "CONNECTION_OWNERSHIP_CONFLICT",
                  "CAPABILITY_NOT_SUPPORTED",
                  "NETWORK_NOT_SUPPORTED",
                  "PROVIDER_UNAVAILABLE",
                  "PROVIDER_RATE_LIMITED",
                  "PROVIDER_REJECTED",
                  "MEDIA_INVALID",
                  "TARGET_INVALID",
                  "TARGET_DROPPED_BY_PROVIDER",
                  "PUBLICATION_FAILED",
                  "PUBLICATION_STATE_UNKNOWN",
                  "REAUTHORIZATION_REQUIRED",
                  "CONNECTION_NOT_ACTIVE",
                  "IDEMPOTENCY_CONFLICT",
                  "IDEMPOTENCY_KEY_REQUIRED",
                  "INVALID_STATE",
                  "VALIDATION_ERROR",
                  "WEBHOOK_SIGNATURE_INVALID",
                  "RATE_LIMITED",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "retryable"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "SocialPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "workspaceId": {
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/SocialContent"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "queued",
              "scheduled",
              "publishing",
              "partially_published",
              "published",
              "failed",
              "cancelled"
            ]
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "timezone": {
            "type": "string"
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialPostTarget"
            }
          },
          "externalRef": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "content",
          "status",
          "targets",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
      "SocialPostPage": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialPost"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/v1/posts": {
      "post": {
        "summary": "Create a post (draft) with its targets",
        "tags": [
          "posts"
        ],
        "description": "Validates every target against the workspace's connections and the network constraints. Idempotent on Idempotency-Key.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequestInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "x-zeptly-workspace-id",
            "required": true,
            "description": "Zeptly workspace identifier (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "description": "REQUIRED (missing → 400 IDEMPOTENCY_KEY_REQUIRED). 8-128 chars [A-Za-z0-9._:-]. Replays return the original response; reuse with a different body → IDEMPOTENCY_CONFLICT."
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialPost"
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialPost"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List posts",
        "tags": [
          "posts"
        ],
        "parameters": [
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "queued",
                "scheduled",
                "publishing",
                "partially_published",
                "published",
                "failed",
                "cancelled"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "x-zeptly-workspace-id",
            "required": true,
            "description": "Zeptly workspace identifier (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialPostPage"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{id}": {
      "get": {
        "summary": "Get a post with per-target status",
        "tags": [
          "posts"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "x-zeptly-workspace-id",
            "required": true,
            "description": "Zeptly workspace identifier (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialPost"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Edit a draft or scheduled post (same post id)",
        "tags": [
          "posts"
        ],
        "description": "Replaces base content and/or the target list (variants, options). Scheduled posts are re-planned: provider copies already handed off are updated in place when supported (OUTSTAND_POST_UPDATE_ENABLED), otherwise replaced. Rejected once any target is publishing or published.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostRequestInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "x-zeptly-workspace-id",
            "required": true,
            "description": "Zeptly workspace identifier (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-caller",
            "required": true,
            "description": "Calling service identity (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-timestamp",
            "required": true,
            "description": "Unix seconds (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-signature",
            "required": true,
            "description": "v1=<hex HMAC-SHA256>"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-zeptly-agent",
            "required": false,
            "description": "Optional agent/user reference recorded in audit events (signed)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "description": "Correlation id; echoed in responses"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "description": "REQUIRED (missing → 400 IDEMPOTENCY_KEY_REQUIRED). 8-128 chars [A-Za-z0-9._:-]. Replays return the original response; reuse with a different body → IDEMPOTENCY_CONFLICT."
          }
        ],
        "security": [
          {
            "zeptlyServiceSignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialPost"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}